refactor(core): isolate local sync persistence
